feat(mac): Added loading screen and reformated files
