+++ mac/Config/Config/InstallationViews/ParentInstallView.swift
@@ -1,18 +1,19 @@
-//
-//  ParentInstallView.swift
-//  Config
-//
-//  Created by Eli Schantz on 6/29/26.
-//
+/*
+ * Keyman is copyright (C) SIL Global. MIT License.
+ *
+ * Created by Eli Schantz on 2026-06-29
+ *
+ * Main view used for configuring Keyman
+ */
 
 import SwiftUI
 
 struct ParentInstallView: View {
   @EnvironmentObject var installation: InstallationContainer
   @Namespace var animation
-  @State public var currentPage: InstallPage = .initial
+  @State public var currentPage: InstallPage = .loading
   
-  func chooseCurrentPage() -> Void {
+  func chooseCurrentPage() {
     if installation.installationPhase.hasTasks {
       switch installation.currentTask()?.taskType {
       case .prepareNewInstall: currentPage = .initial
@@ -20,6 +21,7 @@ struct ParentInstallView: View {
         currentPage = .enableInputMethod
       case .requestAccess: currentPage =
           .allowSecurityPermission
+      case .confirmAccess: print("Access granted")
       case .restartMac: currentPage = .restartComputer
       default: currentPage = .completed
       }
@@ -31,32 +33,44 @@ struct ParentInstallView: View {
       default: currentPage = .completed
       }
     }
+    print("currentPage is now \(currentPage)")
   }
   
   var body: some View {
-    
     ZStack {
       switch currentPage {
-      case .loading: KeymanLogo(namespace: animation)
-      case .initial: InitialView(namespace: animation,onContinue: chooseCurrentPage)
+      case .loading: ProgressView()
+      case .initial: InitialView(namespace: animation,onContinue: {
+        installation.executeNextInstallationTask()
+        chooseCurrentPage()
+      })
       case .completed: CompletedInstallView(namespace: animation)
       case .enableInputMethod: EnableInputMethodView(namespace: animation, onContinue: chooseCurrentPage)
-      case .allowSecurityPermission: CheckAccessibiltyPermissionView(namespace: animation, onContinue: chooseCurrentPage)
+      case .allowSecurityPermission: GrantAccessibiltyPermissionView(namespace: animation, onContinue: chooseCurrentPage)
       case .rerunInstaller: RerunInstallerView(namespace: animation)
       case .restartComputer: RestartComputerView(namespace: animation)
       }
     }
     .onAppear {
-      switch installation.installationPhase {
-      case .evaluatingInstallation:
-        currentPage = .initial
-      default:
+      if installation.installationPhase == .evaluatingInstallation {
+        currentPage = .loading
+        Task {
+          while installation.installationPhase == .evaluatingInstallation {
+            try? await Task.sleep(for: .milliseconds(200))
+          }
+          await MainActor.run {
+            withAnimation(.smooth) {
+              chooseCurrentPage()
+            }
+          }
+        }
+      } else {
         chooseCurrentPage()
       }
     }
     .padding()
-    .frame(width: 600)
-    .frame(height: 400)
+    .frame(minWidth: 600)
+    .frame(minHeight: 400)
   }
 }
