+++ mac/Config/Config/InstallationViews/RestartComputerView.swift
@@ -1,9 +1,10 @@
-//
-//  RestartComputerView.swift
-//  Config
-//
-//  Created by Eli Schantz on 7/21/26.
-//
+/*
+ * Keyman is copyright (C) SIL Global. MIT License.
+ *
+ * Created by Eli Schantz on 2026-07-21
+ *
+ * View used for directing the user to restart their mac.
+ */
 
 import SwiftUI
 
@@ -13,47 +14,29 @@ struct RestartComputerView: View {
   
   var body: some View {
     VStack {
-      
       Text("Restart Computer")
         .font(.title)
         .bold()
         .frame(maxWidth: .infinity, alignment: .center)
         .matchedGeometryEffect(id: "title", in: namespace)
       
-      GradientDivider(namespace: namespace)
-      
       Spacer()
-      
       
       Image(systemName: "restart.circle.fill")
-        .font(.system(size: 88))
-        .padding(8)
-      
-      Text("A restart is required for the previous installation steps to take effect.")
+        .font(.system(size: 100))
+        .padding(16)
+      Text("Restart your Mac to complete the installation. After restarting, open Keyman Configuration again if it doesn't launch automatically.")
         .multilineTextAlignment(.center)
+        .padding(.bottom, 8)
       
       Spacer()
       
+      GradientDivider(namespace: namespace)
+        .padding(.bottom, 8)
       HStack {
-        
         Text("Finish installation")
           .font(.title2)
           .frame(maxWidth: .infinity, alignment: .leading)
-        
-        
-        
-        Button {
-          
-        } label: {
-            Text("Restart Now")
-                .padding(.horizontal, 16)
-                .padding(.vertical, 4)
-        }
-        .buttonStyle(.borderedProminent)
-        .tint(.blue)
-        .clipShape(Capsule())
-        .matchedGeometryEffect(id: "actionButton", in: namespace)
-        
         NavigationButton(action: .dismiss)
       }
     }
