--- mac/Config/Config/InstallerViews/InstallPage.swift
@@ -1,16 +1,0 @@
-//
-//  InstallPage.swift
-//  Config
-//
-//  Created by Eli Schantz on 7/21/26.
-//
-
-enum InstallPage: String, CaseIterable {
-  case loading
-  case initial
-  case completed
-  case enableInputMethod
-  case allowSecurityPermission
-  case rerunInstaller
-  case restartComputer
-}
